+++ assets/javascripts/kitten/components/navigation/header-nav/components/hidden.js
@@ -1,12 +1,17 @@
 import React from 'react'
+import classNames from 'classnames'
 
 export const Hidden = ({ min, max, children }) => {
   const hiddenMin = min ? `k-u-hidden@${min}-up--important` : ''
   const hiddenMax = max ? `k-u-hidden@${max}-down--important` : ''
 
   return (
-    <div className={`k-HeaderNav--hidden ${hiddenMin} ${hiddenMax}`.trim()}>
-      {children}
-    </div>
+    <>
+      {React.Children.map(children, child =>
+        React.cloneElement(child, {
+          className: classNames(hiddenMin, hiddenMax),
+        }),
+      )}
+    </>
   )
 }